pyproject.toml
[tool.poetry]
name = "cw-7"
version = "0.1.0"
description = ""
authors = ["Dmitry Zadorozhnyy <zadorozhnyy_d_m@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
django-stubs = "^5.1.1"
ipython = "^8.29.0"
django-filter = "^24.3"
coverage = "^7.6.7"
celery = "^5.4.0"
redis = "^5.2.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.6.0"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
